refactor steps to no longer use hard-coded ids

+++ xivo_steps/restapi_1_1/users_steps.py
@@ -34,7 +34,8 @@ def _create_user(userinfo):
 
 @step(u'Given there are no users with id "([^"]*)"$')
 def given_there_are_no_users_with_id_group1(step, user_id):
-    user_helper.delete_user(user_id)
+    if user_helper.user_exists(user_id):
+        user_helper.delete_user(user_id)
 
 
 @step(u'When I ask for the list of users$')
@@ -42,9 +43,15 @@ def when_i_ask_for_the_list_of_users(step):
     world.response = user_action_restapi.all_users()
 
 
-@step(u'When I ask for the user with id "([^"]*)"$')
-def when_i_ask_for_the_user_with_id_group1(step, userid):
-    world.response = user_action_restapi.get_user(userid)
+@step(u'When I ask for a fake user$')
+def when_i_ask_for_a_fake_user(step):
+    world.response = user_action_restapi.get_user(999999999)
+
+
+@step(u'When I ask for the user with name "([^"]*)" "([^"]*)"$')
+def when_i_ask_for_the_user_with_id_group1(step, firstname, lastname):
+    user = user_helper.get_by_firstname_lastname(firstname, lastname)
+    world.response = user_action_restapi.get_user(user['id'])
 
 
 @step(u'When I search for the user "([^"]*)"$')
@@ -63,14 +70,13 @@ def when_i_create_users_with_the_following_parameters(step):
         world.response = user_action_restapi.create_user(userinfo)
 
 
-@step(u'When I update the user with id "([^"]*)" using the following parameters:$')
-def when_i_update_the_user_with_id_group1_using_the_following_parameters(step, userid):
-    userinfo = _get_user_info(step.hashes)
-    world.response = user_action_restapi.update_user(userid, userinfo)
+@step(u'When I update a fake user')
+def when_i_edit_a_fake_user(step):
+    world.response = user_action_restapi.update_user(999999999, {})
 
 
-@step(u'When I update user "([^"]*)" "([^"]*)" with the following parameters:')
-def when_i_update_user_group1_group2_with_the_following_parameters(step, firstname, lastname):
+@step(u'When I update the user with name "([^"]*)" "([^"]*)" using the following parameters:$')
+def when_i_update_the_user_with_id_group1_using_the_following_parameters(step, firstname, lastname):
     user = user_helper.get_by_firstname_lastname(firstname, lastname)
     userinfo = _get_user_info(step.hashes)
     world.response = user_action_restapi.update_user(user['id'], userinfo)
@@ -85,6 +91,11 @@ def when_i_delete_the_user_with_id_group1(step, userid):
 def when_i_delete_the_user_with_name_group1_group2(step, firstname, lastname):
     user = user_helper.get_by_firstname_lastname(firstname, lastname)
     world.response = user_action_restapi.delete_user(user['id'])
+
+
+@step(u'When I delete a fake user')
+def when_i_delete_a_fake_user(step):
+    world.response = user_action_restapi.delete_user(999999999)
 
 
 @step(u'Then I get a list containing the following users:')
@@ -126,10 +137,10 @@ def then_the_created_user_has_the_following_parameters(step):
     assert_that(user, has_entries(expected_user))
 
 
-@step(u'Then the user with id "([^"]*)" no longer exists')
-def then_the_user_with_id_group1_no_longer_exists(step, userid):
-    response = user_action_restapi.get_user(userid)
-    assert_that(response.status, equal_to(404))
+@step(u'Then the user with name "([^"]*)" "([^"]*)" no longer exists')
+def then_the_user_with_id_group1_no_longer_exists(step, firstname, lastname):
+    user = user_helper.find_by_firstname_lastname(firstname, lastname)
+    assert_that(user, none())
 
 
 @step(u'Then I get a response with a user id')
